adding cloudshell option

+++ Common.ps1
@@ -12,18 +12,26 @@ if ($headers) {
 
 Try {
     # Install latest AD client library
-    $nuget = "nuget.exe"
-    if (!(test-path $nuget)) {
-        $nugetDownloadUrl = "https://dist.nuget.org/win-x86-commandline/latest/nuget.exe"
-        invoke-webRequest $nugetDownloadUrl -outFile $nuget
+    # check for cloudshell 
+    $cloudShellDll = '/lib/azure-functions-core-tools/Microsoft.IdentityModel.Clients.ActiveDirectory.dll'
+    
+    if ($PSVersionTable.Platform -ieq 'unix' -and (test-path $cloudShellDll)) {
+        $FilePath = $cloudShellDll
     }
+    else {
+        $nuget = "nuget.exe"
+        if (!(test-path $nuget)) {
+            $nugetDownloadUrl = "https://dist.nuget.org/win-x86-commandline/latest/nuget.exe"
+            invoke-webRequest $nugetDownloadUrl -outFile $nuget
+        }
 
-    $env:path = $env:path.replace(";$pwd;", "") + ";$pwd;"
-    $ADPackage = "Microsoft.IdentityModel.Clients.ActiveDirectory"
-    & nuget.exe install $ADPackage > nuget.log
+        $env:path = $env:path.replace(";$pwd;", "") + ";$pwd;"
+        $ADPackage = "Microsoft.IdentityModel.Clients.ActiveDirectory"
+        & nuget.exe install $ADPackage > nuget.log
 
-    # Target .NET Framework version of the DLL
-    $FilePath = (Get-Item .\\Microsoft.IdentityModel.Clients.ActiveDirectory.[0-9].[0-9].[0-9]\\lib\\net[0-9][0-9]\\Microsoft.IdentityModel.Clients.ActiveDirectory.dll).FullName | Resolve-Path -Relative
+        # Target .NET Framework version of the DLL
+        $FilePath = (Get-Item .\\Microsoft.IdentityModel.Clients.ActiveDirectory.[0-9].[0-9].[0-9]\\lib\\net[0-9][0-9]\\Microsoft.IdentityModel.Clients.ActiveDirectory.dll).FullName | Resolve-Path -Relative
+    }
     Add-Type -Path $FilePath
 }
 Catch {
